+++ binance/src/model/websocket.rs
@@ -1,11 +1,13 @@
 use serde::{Deserialize, Serialize};
 
-use shared::string_to_float;
+use shared::string_to_decimal;
 
 use super::{
     Asks, Bids, Kline, OrderBook, OrderExecType, OrderRejectReason, OrderStatus, OrderType, Side,
     TimeInForce,
 };
+
+use rust_decimal::prelude::Decimal;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum Subscription {
@@ -51,10 +53,10 @@ pub struct TradeMessage {
     pub symbol: String,
     #[serde(rename = "t")]
     pub trade_id: u64,
-    #[serde(rename = "p", with = "string_to_float")]
-    pub price: f64,
-    #[serde(rename = "q", with = "string_to_float")]
-    pub qty: f64,
+    #[serde(rename = "p", with = "string_to_decimal")]
+    pub price: Decimal,
+    #[serde(rename = "q", with = "string_to_decimal")]
+    pub qty: Decimal,
     #[serde(rename = "b")]
     pub buyer_order_id: u64,
     #[serde(rename = "a")]
@@ -78,10 +80,10 @@ pub struct AggregateTrade {
     pub symbol: String,
     #[serde(rename = "a")]
     pub aggregated_trade_id: u64,
-    #[serde(rename = "p", with = "string_to_float")]
-    pub price: f64,
-    #[serde(rename = "q", with = "string_to_float")]
-    pub qty: f64,
+    #[serde(rename = "p", with = "string_to_decimal")]
+    pub price: Decimal,
+    #[serde(rename = "q", with = "string_to_decimal")]
+    pub qty: Decimal,
     #[serde(rename = "f")]
     pub first_break_trade_id: u64,
     #[serde(rename = "l")]
@@ -111,14 +113,14 @@ pub struct UserOrderUpdate {
     pub order_type: OrderType,
     #[serde(rename = "f")]
     pub time_in_force: TimeInForce,
-    #[serde(rename = "q", with = "string_to_float")]
-    pub qty: f64,
-    #[serde(rename = "p", with = "string_to_float")]
-    pub price: f64,
-    #[serde(rename = "P", with = "string_to_float")]
-    pub stop_price: f64,
-    #[serde(rename = "F", with = "string_to_float")]
-    pub iceberg_qty: f64,
+    #[serde(rename = "q", with = "string_to_decimal")]
+    pub qty: Decimal,
+    #[serde(rename = "p", with = "string_to_decimal")]
+    pub price: Decimal,
+    #[serde(rename = "P", with = "string_to_decimal")]
+    pub stop_price: Decimal,
+    #[serde(rename = "F", with = "string_to_decimal")]
+    pub iceberg_qty: Decimal,
     #[serde(skip_serializing)]
     pub g: i32,
     #[serde(skip_serializing, rename = "C")]
@@ -131,14 +133,14 @@ pub struct UserOrderUpdate {
     pub order_reject_reason: OrderRejectReason,
     #[serde(rename = "i")]
     pub order_id: u64,
-    #[serde(rename = "l", with = "string_to_float")]
-    pub qty_last_filled_trade: f64,
-    #[serde(rename = "z", with = "string_to_float")]
-    pub accumulated_qty_filled_trades: f64,
-    #[serde(rename = "L", with = "string_to_float")]
-    pub price_last_filled_trade: f64,
-    #[serde(rename = "n", with = "string_to_float")]
-    pub commission: f64,
+    #[serde(rename = "l", with = "string_to_decimal")]
+    pub qty_last_filled_trade: Decimal,
+    #[serde(rename = "z", with = "string_to_decimal")]
+    pub accumulated_qty_filled_trades: Decimal,
+    #[serde(rename = "L", with = "string_to_decimal")]
+    pub price_last_filled_trade: Decimal,
+    #[serde(rename = "n", with = "string_to_decimal")]
+    pub commission: Decimal,
     #[serde(skip_serializing, rename = "N")]
     pub asset_commisioned: Option<String>,
     #[serde(rename = "T")]
@@ -155,8 +157,8 @@ pub struct UserOrderUpdate {
     pub m_ignore: bool,
     #[serde(skip_serializing, rename = "O")]
     pub order_creation_time: u64,
-    #[serde(skip_serializing, rename = "Z", with = "string_to_float")]
-    pub cumulative_quote_asset_transacted_qty: f64,
+    #[serde(skip_serializing, rename = "Z", with = "string_to_decimal")]
+    pub cumulative_quote_asset_transacted_qty: Decimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -187,36 +189,36 @@ pub struct Ticker {
     pub event_time: u64,
     #[serde(rename = "s")]
     pub symbol: String,
-    #[serde(rename = "p", with = "string_to_float")]
-    pub price_change: f64,
-    #[serde(rename = "P", with = "string_to_float")]
-    pub price_change_percent: f64,
-    #[serde(rename = "w", with = "string_to_float")]
-    pub average_price: f64,
-    #[serde(rename = "x", with = "string_to_float")]
-    pub prev_close: f64,
-    #[serde(rename = "c", with = "string_to_float")]
-    pub current_close: f64,
-    #[serde(rename = "Q", with = "string_to_float")]
-    pub current_close_qty: f64,
-    #[serde(rename = "b", with = "string_to_float")]
-    pub best_bid: f64,
-    #[serde(rename = "B", with = "string_to_float")]
-    pub best_bid_qty: f64,
-    #[serde(rename = "a", with = "string_to_float")]
-    pub best_ask: f64,
-    #[serde(rename = "A", with = "string_to_float")]
-    pub best_ask_qty: f64,
-    #[serde(rename = "o", with = "string_to_float")]
-    pub open: f64,
-    #[serde(rename = "h", with = "string_to_float")]
-    pub high: f64,
-    #[serde(rename = "l", with = "string_to_float")]
-    pub low: f64,
-    #[serde(rename = "v", with = "string_to_float")]
-    pub volume: f64,
-    #[serde(rename = "q", with = "string_to_float")]
-    pub quote_volume: f64,
+    #[serde(rename = "p", with = "string_to_decimal")]
+    pub price_change: Decimal,
+    #[serde(rename = "P", with = "string_to_decimal")]
+    pub price_change_percent: Decimal,
+    #[serde(rename = "w", with = "string_to_decimal")]
+    pub average_price: Decimal,
+    #[serde(rename = "x", with = "string_to_decimal")]
+    pub prev_close: Decimal,
+    #[serde(rename = "c", with = "string_to_decimal")]
+    pub current_close: Decimal,
+    #[serde(rename = "Q", with = "string_to_decimal")]
+    pub current_close_qty: Decimal,
+    #[serde(rename = "b", with = "string_to_decimal")]
+    pub best_bid: Decimal,
+    #[serde(rename = "B", with = "string_to_decimal")]
+    pub best_bid_qty: Decimal,
+    #[serde(rename = "a", with = "string_to_decimal")]
+    pub best_ask: Decimal,
+    #[serde(rename = "A", with = "string_to_decimal")]
+    pub best_ask_qty: Decimal,
+    #[serde(rename = "o", with = "string_to_decimal")]
+    pub open: Decimal,
+    #[serde(rename = "h", with = "string_to_decimal")]
+    pub high: Decimal,
+    #[serde(rename = "l", with = "string_to_decimal")]
+    pub low: Decimal,
+    #[serde(rename = "v", with = "string_to_decimal")]
+    pub volume: Decimal,
+    #[serde(rename = "q", with = "string_to_decimal")]
+    pub quote_volume: Decimal,
     #[serde(rename = "O")]
     pub open_time: u64,
     #[serde(rename = "C")]
@@ -274,10 +276,10 @@ pub struct AccountUpdate {
 pub struct AccountUpdateBalance {
     #[serde(rename = "a")]
     pub asset: String,
-    #[serde(rename = "f", with = "string_to_float")]
-    pub free: f64,
-    #[serde(rename = "l", with = "string_to_float")]
-    pub locked: f64,
+    #[serde(rename = "f", with = "string_to_decimal")]
+    pub free: Decimal,
+    #[serde(rename = "l", with = "string_to_decimal")]
+    pub locked: Decimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -289,16 +291,16 @@ pub struct MiniTicker {
     pub event_time: u64,
     #[serde(rename = "s")]
     pub symbol: String,
-    #[serde(rename = "c", with = "string_to_float")]
-    pub close: f64,
-    #[serde(rename = "o", with = "string_to_float")]
-    pub open: f64,
-    #[serde(rename = "l", with = "string_to_float")]
-    pub low: f64,
-    #[serde(rename = "h", with = "string_to_float")]
-    pub high: f64,
-    #[serde(rename = "v", with = "string_to_float")]
-    pub volume: f64,
-    #[serde(rename = "q", with = "string_to_float")]
-    pub quote_volume: f64,
+    #[serde(rename = "c", with = "string_to_decimal")]
+    pub close: Decimal,
+    #[serde(rename = "o", with = "string_to_decimal")]
+    pub open: Decimal,
+    #[serde(rename = "l", with = "string_to_decimal")]
+    pub low: Decimal,
+    #[serde(rename = "h", with = "string_to_decimal")]
+    pub high: Decimal,
+    #[serde(rename = "v", with = "string_to_decimal")]
+    pub volume: Decimal,
+    #[serde(rename = "q", with = "string_to_decimal")]
+    pub quote_volume: Decimal,
 }
